Fixup: QL to stratagem, pool names in config struct
move hotpool QL to stratagem.rs
Use pool names in HotpoolConfiguration struct.
